Add DigitalCredentials handling impl. example.
The DocumentManagerImpl Builder was updated to demonstrate the possible setup of Digital Credentials support within the DocumentManagerImpl init constructor to facilitate the setup during the Wallet application initialization. However, for the real implementation the dependency graph can be used to supply actual document types supported by the Wallet app or they could be passes as a parameter to a public method of the the DocumentManager or its builder.

The multipaz dependency version bumped to 0.93 to support Digital Credentials.

The Clock replaced with kotlin.time.Clock as it's shown as deprecation with error